chore: move layout components out of shared
