+++ packages/apps/human-app/frontend/src/router/components/drawer-menu-items/drawer-menu-items-worker.tsx
@@ -2,7 +2,7 @@ import { t } from 'i18next';
 import type {
   BottomMenuItem,
   TopMenuItem,
-} from '@/shared/components/layout/protected/drawer-navigation';
+} from '@/router/components/layout/protected/drawer-navigation';
 import {
   HelpIcon,
   UserOutlinedIcon,
