+++ packages/apps/human-app/frontend/src/router/components/layout/protected/layout.tsx
@@ -4,12 +4,12 @@ import { useEffect, useRef, useState } from 'react';
 import { Outlet } from 'react-router-dom';
 import { useIsMobile } from '@/shared/hooks/use-is-mobile';
 import { useBackgroundContext } from '@/shared/contexts/background';
-import type { PageHeaderProps } from '@/shared/components/layout/protected/page-header';
-import { PageHeader } from '@/shared/components/layout/protected/page-header';
+import type { PageHeaderProps } from '@/router/components/layout/protected/page-header';
+import { PageHeader } from '@/router/components/layout/protected/page-header';
 import { breakpoints } from '@/shared/styles/breakpoints';
 import { useIsHCaptchaLabelingPage } from '@/shared/hooks/use-is-hcaptcha-labeling-page';
 import { GovernanceBanner } from '@/modules/governance-banner/components/governance-banner';
-import { Footer } from '../footer';
+import { Footer } from '../../footer';
 import { Navbar } from './navbar';
 
 const Main = styled('main', {
@@ -39,7 +39,7 @@ export function Layout({
   renderDrawer,
   renderHCaptchaStatisticsDrawer,
   renderGovernanceBanner,
-}: {
+}: Readonly<{
   pageHeaderProps: PageHeaderProps;
   renderDrawer: (
     open: boolean,
@@ -47,7 +47,7 @@ export function Layout({
   ) => JSX.Element;
   renderHCaptchaStatisticsDrawer?: (isOpen: boolean) => JSX.Element;
   renderGovernanceBanner?: boolean;
-}) {
+}>) {
   const layoutElementRef = useRef<HTMLDivElement>();
   const isHCaptchaLabelingPage = useIsHCaptchaLabelingPage();
   const isMobile = useIsMobile();
